pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gitnarrative"
version = "0.1.0"
description = "LLM-powered CLI that reads git history and generates narrative markdown describing a project's evolution"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "anthropic>=0.40",
    "jinja2>=3.1",
    "gitpython>=3.1",
]

[project.scripts]
gitnarrative = "gitnarrative.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["gitnarrative"]
